refactor: module and package with same name is bad
